Moved HasReferenceElement to kbl2.4 and kbl2.5
