+++ kbl/kbl-v24/src/main/resources/kbl24/KBL24_SR1-ext.xjb
@@ -86,7 +86,7 @@
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Reference_element']]">
-        <inheritance:implements>com.foursoft.harness.kbl.common.HasReferenceElement&lt;com.foursoft.harness.kbl.v24.ReferenceElement&gt;</inheritance:implements>
+        <inheritance:implements>com.foursoft.harness.kbl.v24.HasReferenceElement</inheritance:implements>
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Description']]">
